lecture_1/pyproject.toml
[project]
name = "lecture-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.0",
]
